requirements.txt
streamlit
spacy
transformers
torch
pdfminer.six
python-docx
scikit-learn
nltk
pandas
